Mention already implemented commands missing from README
There were only two commands missing from the README that had been
implemented but not yet mentioned in the examples section.  This commit adds
a quick example of both so that users can be more aware of the available
functions without having to search in perl5's Shell::Command, which itself
redirects to ExtUtils::Command for a list of the implemented commands.